Extract StateIdAddressPlugin module
Pull out all the logic into an include-able module so all you have to provide is a proofer + cost token.
